+++ src/bin/pod_agent.adb
@@ -3,14 +3,14 @@
 
 with Ada.Strings.Unbounded;
 with Podmander.Agent;
-with Podmander.CLI;
+with Podmander.Args;
 with Podmander.Logging;
 
 procedure Pod_Agent is
    use Ada.Strings.Unbounded;
 begin
    declare
-      Level_Str : constant String := Podmander.CLI.Get ("log-level", "info");
+      Level_Str : constant String := Podmander.Args.Get ("log-level", "info");
    begin
       Podmander.Logging.Set_Level
         (Podmander.Logging.Log_Level'Value (Level_Str));
@@ -22,10 +22,10 @@ begin
    end;
 
    declare
-      Token : constant String := Podmander.CLI.Get ("token", "");
+      Token : constant String := Podmander.Args.Get ("token", "");
    begin
       if Token = "" then
-         Podmander.CLI.Print_Usage
+         Podmander.Args.Print_Usage
            ("pod_agent --token <TOKEN> [--connect <ADDR>] "
             & "[--name <NAME>] [--interval <SEC>] [--log-level <LEVEL>]");
          return;
@@ -36,12 +36,12 @@ begin
       Config : constant Podmander.Agent.Agent_Config :=
         (Controller_Address   =>
            To_Unbounded_String
-             (Podmander.CLI.Get ("connect", "tcp://localhost:5555")),
+             (Podmander.Args.Get ("connect", "tcp://localhost:5555")),
          Agent_Name           =>
-           To_Unbounded_String (Podmander.CLI.Get ("name", "agent-1")),
+           To_Unbounded_String (Podmander.Args.Get ("name", "agent-1")),
          Join_Token           =>
-           To_Unbounded_String (Podmander.CLI.Get ("token", "")),
-         Heartbeat_Interval   => Podmander.CLI.Get_Duration ("interval", 30.0),
+           To_Unbounded_String (Podmander.Args.Get ("token", "")),
+         Heartbeat_Interval   => Podmander.Args.Get_Duration ("interval", 30.0),
          Registration_Timeout => 5.0,
          Max_Backoff          => 60.0);
    begin